+++ src/gGfx/gGfx.h
@@ -33,7 +33,7 @@ public:
 };
 
 
-#define KEY_NUMBERS 256		
+#define KEY_NUMBERS 256
 #define MOUSE_BUTTONS 5
 
 enum class DIRECTION_2D
@@ -219,10 +219,10 @@ public:
 private:
     wchar_t* glyph = nullptr;
     short* colour = nullptr;
-    gint width;
-    gint height;
+    size_t width;
+    size_t height;
 
-    void init(gint w, gint h);
+    void init(size_t w, size_t h);
 
 };
